+++ changelog.md
@@ -5,6 +5,7 @@ __1.14.0__
 * PatchMatchFilter algorithm for image editing, inpainting and completion, in progress
 * ColorDetectorFilter variation of ConnectedComponentsFilter
 * FILTER.Util.Image.Selection class, FILTER.Util.Image.Pyramid class
+* FILTER.Image.select method can accept FILTER.Util.Image.Selection as well
 * handle metaData serialization in CompositeFilter
 
 __1.13.0__